Reorganise the ETL folder by pipeline role — promote helpers/places/ to places/, split helpers.ts into writers/, util/, and   places/cache.ts, and move the ingestor base into ingest/ — a pure relocation with imports repointed and helpers/ deleted.
