+++ packages/db/src/etl/places/place-inference.ts
@@ -1,6 +1,6 @@
-import { db } from '../../../client';
-import { createCachedResolver } from '../helpers';
-import { PlaceIdRow } from '../../../row-types';
+import { db } from '../../client';
+import { createCachedResolver } from './cache';
+import { PlaceIdRow } from '../../row-types';
 import { sql } from 'drizzle-orm/sql';
 import { getCandidatesByPoint, getCandidatesByName, pickFinest, resolveNamePool } from './place-candidates';
 import type { Resolved } from './place-index';
