+++ packages/db/src/etl/sources/adressen.ts
@@ -18,7 +18,7 @@ import { readFileSync } from 'fs';
 import { Parser } from 'n3';
 import { sql } from 'drizzle-orm';
 import { db } from '../../client';
-import { createNameWriter } from '../helpers/helpers';
+import { createNameWriter } from '../writers/place-writer';
 
 export async function ingest(filePath: string) {
   console.log(`Parsing ${filePath}...`);
