+++ packages/db/src/etl/sources/lps.ts
@@ -12,7 +12,7 @@
  */
 import { readFileSync } from 'fs';
 import { Parser } from 'n3';
-import { insertPlaces, type PlaceInsert } from '../helpers/helpers';
+import { insertPlaces, type PlaceInsert } from '../writers/place-writer';
 
 export async function ingest(filePath: string) {
   console.log(`Parsing ${filePath}...`);
